added protected route and middleware
